Import and implement server module for web server startup
Introduces a new `server` module to manage the startup of a web server. It replaces the previous command-line argument parsing and server binding logic with a new `start_server` function within the `server` module
