+++ src/main.rs
@@ -4,6 +4,7 @@ mod database;
 mod models;
 mod output_formatter;
 mod utils;
+mod server; // Added this line
 
 /// The main entrypoint of the application
 fn main() {
